requirements.txt
fastapi==0.61.1
bcrypt==3.2.0
uvicorn==0.12.2
python-jose==3.2.0
passlib==1.7.4
aiocache==0.11.1
aiohttp==3.7.3
aiohttp-retry==2.3.3
ujson==5.2.0

python-multipart==0.0.5
odmantic==0.3.5
pydantic==1.8.2
email-validator==1.1.1
aiosmtplib==1.1.6
motor==2.3.0
transliterate==1.10.2
Pillow==9.1.0
